Adding authorize commands

+++ Packages/CommandHandler/command-handler/server_component.go
@@ -18,30 +18,48 @@ func routeServerAction(component gu.DiscordComponentInteraction) (data gu.Discor
 
 	}
 
-	var message string
-	switch action {
-	case "start":
-		log.Printf("Starting server")
-		err = server.Start()
-		if err == nil {
-			message = "Starting server"
-		}
-	case "stop":
-		log.Printf("Sopping server")
-		err = server.Stop()
-		if err == nil {
-			message = "Stopping server"
-		}
-	case "reboot":
-		log.Printf("Rebooting server")
-		err = server.Restart()
-		if err == nil {
-			message = "Rebooting server"
+	var authorized bool
+	for _, user := range server.AuthorizedUsers() {
+		if user == component.Member.User.ID {
+			authorized = true
 		}
 	}
-	if err != nil {
-		log.Printf("Error performing command: %v", err)
-		message = fmt.Sprintf("Error running %v on server", action)
+	for _, role := range server.AuthorizedRoles() {
+		for _, userRole := range component.Member.Roles {
+			if role == userRole {
+				authorized = true
+			}
+		}
+	}
+
+	var message string
+	if authorized {
+		switch action {
+		case "start":
+			log.Printf("Starting server")
+			err = server.Start()
+			if err == nil {
+				message = "Starting server"
+			}
+		case "stop":
+			log.Printf("Sopping server")
+			err = server.Stop()
+			if err == nil {
+				message = "Stopping server"
+			}
+		case "reboot":
+			log.Printf("Rebooting server")
+			err = server.Restart()
+			if err == nil {
+				message = "Rebooting server"
+			}
+		}
+		if err != nil {
+			log.Printf("Error performing command: %v", err)
+			message = fmt.Sprintf("Error running %v on server", action)
+		}
+	} else {
+		message = "You do not have authorization to run commands on this server."
 	}
 	log.Printf("Message to send to discord")
 	data = gu.DiscordInteractionResponseData{
